% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\title{Backward}
\usage{
backward(example, switching, visible, init)
}
\arguments{
\item{example}{example HMM path to be compared to}

\item{switching}{probability of each hidden state at the next time step}

\item{visible}{probability of hidden state producing a visible state}

\item{init}{initial probability for each hidden state (at time step 1)}
}
\value{
A matrix named example_p that contains the probability of each hidden state at each time step.
}
\description{
This function finds the probability of given model producing the
same result as the example set of states from a given endpoint.
Basically the reverse of forward.
}
