% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{Viterbi}
\alias{Viterbi}
\title{Viterbi}
\usage{
Viterbi(example, switching, visible, initial_distribution)
}
\arguments{
\item{example}{An example HMM, on which various statistics are derived}

\item{switching}{the probability of switching hidden states (A->A, ..., B->A)}

\item{visible}{the probability of a hidden state resulting in a visible state.}

\item{initial_distribution}{probability of the initial hidden state.}
}
\value{
A list with the most likely path of hidden states through the HMM
}
\description{
Function to path trace the most likely path through a Hidden
Markov Model given an example model and an end state.
}
